Replaced old shard_creator with new version
